+++ HISTORY.md
@@ -2,10 +2,11 @@ History
 ===============
 
 ### 0.0.3 2013-09-
-* 1 major enhacement
+* 2 major enhacement
 	*	Bumped Bioinfo::Database::HGNC version to 0.2.0
   	* Used metaprogramming in Bioinfo::Database::HGNC
   	* Removd bioinfo/databases/hgnc/*
+  * Added Bioinfo::Scripts::Cipher
 
 ### 0.0.2 2013-09-03
 * 2 major enhancement